Task done
The task implements an API connection using async await. We receive information from the user about which movie he wants to find on the service. After clicking on the "Search" button, a request is sent in order to obtain information about the films.

For each Details button, an id attribute with a unique code on the site will be assigned to facilitate further search on the site.

After clicking on one of the Details buttons, full information about the movie is displayed.